+++ README.md
@@ -5,3 +5,13 @@
 ## 🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄🎁🎄
 
 ## Solutions for Advent of Code puzzles by [Domen Gabrovšek](https://www.github.com/domengabrovsek).
+
+## Get started
+
+```js
+// init template source file for specified year
+~ npm run init [year] // npm run init 2022
+
+// run solution for specific year and day (inputs are fetched and saved automatically)
+~ npm run [year] [day] // npm run 2022 1
+```
