+++ src/2018/solutions/day-23.ts
@@ -1,7 +1,6 @@
 /* This file contains solution for AoC puzzle day 23 */
 
-import { getInput } from "../../utils/inputs";
-import { getYear, getDay } from '../../utils/utils';
+import { getYear, getDay, getInput } from '../../utils/utils';
 
 export const solveOne = async () => {
